rexus/requirements.txt
# Rexus.app Dependencies - Updated with improvements

# Core Framework
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Database
sqlite3  # Built-in Python module

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Performance Monitoring
psutil>=5.9.0

# Development and Testing
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Optional Web Components (graceful degradation)
PyQt6-WebEngine>=6.6.0  # Optional for web features

# Logging and Monitoring
colorlog>=6.7.0  # For colored console logs

# Configuration
python-dotenv>=1.0.0  # For .env file support

# Data Processing
pandas>=2.0.0  # For data export/import features
openpyxl>=3.1.0  # For Excel file support

# Utilities
pathlib  # Built-in Python module
typing-extensions>=4.7.0

# Optimizations
lru-cache>=0.1.0  # Cache LRU adicional
memory-profiler>=0.60.0  # Profiling de memoria
